+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * _add_ attack module adds projectile entities to the engine
 * _add_ ship components
 * _add_ thrust and hull components
+* _add_ ship hardpoints for mounting components
 
 # 0.0.4